05_vectorize/export_postgis.py
"""
PostGIS Export Module

Exports vectorized and simplified polygons to the PostGIS database
in the dev_rcxl.predicted_polygons table.

Will be implemented in Sprint 4.
"""

import logging
from typing import Dict, List, Optional

from shapely.geometry import MultiPolygon
import psycopg2

import sys
from pathlib import Path
sys.path.insert(0, str(Path(__file__).parent.parent))
from config import Config


logger = logging.getLogger(__name__)


def export_predictions_to_postgis(
    conn,
    chart_id: int,
    model_version: str,
    polygons: Dict[str, MultiPolygon],
    confidence_scores: Dict[str, float],
    simplify_tolerance: float
) -> List[int]:
    """
    Export prediction polygons to PostGIS database.
    
    Args:
        conn: Database connection
        chart_id: Chart ID
        model_version: Model version identifier
        polygons: Dictionary mapping class types to MultiPolygons
        confidence_scores: Dictionary mapping class types to confidence scores
        simplify_tolerance: Tolerance used for simplification
        
    Returns:
        List of inserted prediction IDs
        
    Raises:
        NotImplementedError: Will be implemented in Sprint 4
    """
    raise NotImplementedError("Will be implemented in Sprint 4")


def insert_prediction_polygon(
    conn,
    chart_id: int,
    model_version: str,
    class_type: str,
    geom: MultiPolygon,
    confidence_mean: float,
    pixel_area: int,
    simplify_tolerance: float
) -> Optional[int]:
    """
    Insert a single prediction polygon into database.
    
    Args:
        conn: Database connection
        chart_id: Chart ID
        model_version: Model version
        class_type: Class type ('sea', 'land', 'exclude')
        geom: MultiPolygon geometry
        confidence_mean: Mean confidence score
        pixel_area: Area in pixels
        simplify_tolerance: Simplification tolerance used
        
    Returns:
        Prediction ID or None if failed
        
    Raises:
        NotImplementedError: Will be implemented in Sprint 4
    """
    raise NotImplementedError("Will be implemented in Sprint 4")


def update_chart_status(
    conn,
    chart_id: int,
    status: str = 'predicted'
):
    """
    Update chart status after prediction export.
    
    Args:
        conn: Database connection
        chart_id: Chart ID
        status: New status
        
    Raises:
        NotImplementedError: Will be implemented in Sprint 4
    """
    raise NotImplementedError("Will be implemented in Sprint 4")


def verify_complete_coverage(
    conn,
    chart_id: int,
    pred_ids: List[int]
) -> bool:
    """
    Verify that predictions provide complete coverage.
    
    Checks that predicted polygons cover the entire chart area.
    
    Args:
        conn: Database connection
        chart_id: Chart ID
        pred_ids: List of prediction IDs
        
    Returns:
        True if complete coverage is achieved
        
    Raises:
        NotImplementedError: Will be implemented in Sprint 4
    """
    raise NotImplementedError("Will be implemented in Sprint 4")
