requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0
rtree>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0

# Image processing
scikit-image>=0.21.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
click>=8.1.0

# Testing
pytest>=7.4.0

# GDAL (typically installed at system level, but included for reference)
# GDAL>=3.6.0
